--- index.html
@@ -1,54 +1,0 @@
-
-<html>
-    <head>
-    <title>Chat Room</title>
-    <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/1.4.8/socket.io.min.js"></script>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/2.2.4/jquery.min.js"></script>
-    </head>
-    <body>
-    <script type="text/javascript">
-    $(document).ready(function() {
-    
-        var socket = io.connect('http://127.0.0.1:5000');
-        var socket_attendence = io('http://127.0.0.1:5000/attendence_namespace')
-        var socket_admin = io('http://127.0.0.1:5000/admin_namespace')
-        var socket_rooms = io('http://127.0.0.1:5000/rooms_namespace')
-    
-        $('#attendence_button').on('click', function() {
-            socket_attendence.emit('attendence_request', {'otp':$('#myMessage').val(), 'token':'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJleHAiOjE2MTE2Mzc2NDksImlhdCI6MTU4MDEwMTY0NCwic3ViIjoxfQ.QkM-DsJR0yuVbSL-ElkQeHG0xYPpF_BrA-lJhi8MZqk'});
-            $('#myMessage').val('');
-        });
-
-        socket_attendence.on('attendence_result', function(json) {
-		$("#messages").append('<li>'+json+'</li>');
-		console.log('Received message on attendence_namespace');
-	});
-
-
-
-    socket_admin.on('admin_listen', function(json) {
-		$("#messages").append('<li>'+json+'</li>');
-        $("#messages").append('<li>'+'upper one recived for admin_namespace'+'</li>');
-        // Check if it's allowed to broadcast to rooms, if not then don't emit below to room socket else do emit
-        socket_rooms.emit('join', json);
-		console.log('Received message on admin_namespace');
-	});
-
-    socket_rooms.on('join_room', function(msg) {
-        console.log(msg);
-    });
-
-    socket_rooms.on('leave_room', function(msg) {
-        console.log(json);
-    });
-
-
-});
-      
-      </script>
-    <ul id="messages"></ul>
-    <input type="text" id="myMessage">
-    <!-- <button id="sendbutton">Send</button> -->
-    <button id="attendence_button">Attendence</button>
-    </body>
-    </html>
